Tiny refactoring
