+++ src/XamlNameReferenceGenerator.Sandbox/Program.cs
@@ -3,7 +3,7 @@ using Avalonia.ReactiveUI;
 
 namespace XamlNameReferenceGenerator.Sandbox
 {
-    class Program
+    internal static class Program
     {
         public static void Main(string[] args) => BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
