+++ src/XamlNameReferenceGenerator.Sandbox/Views/SignUpView.xaml
@@ -1,8 +1,7 @@
 ﻿<Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:sandbox="clr-namespace:XamlNameReferenceGenerator.Sandbox"
         xmlns:controls="clr-namespace:XamlNameReferenceGenerator.Sandbox.Controls"
-        x:Class="XamlNameReferenceGenerator.Sandbox.SignUpView">
+        x:Class="XamlNameReferenceGenerator.Sandbox.Views.SignUpView">
     <StackPanel>
         <!-- Regular name directives. -->
         <controls:CustomTextBox Margin="0 10 0 0"
@@ -21,6 +20,7 @@
                    Foreground="Red"
                    FontSize="12" />
 
+        <!-- DataTemplates should be removed. -->
         <ListBox x:Name="AwesomeListView">
             <ListBox.ItemTemplate>
                 <DataTemplate>
