+++ src/XamlNameReferenceGenerator.Tests/Views/CustomControls.xml
