+++ src/XamlNameReferenceGenerator.Tests/Views/DataTemplates.xml
