+++ src/XamlNameReferenceGenerator.Tests/Views/NamedControl.xml
