+++ src/XamlNameReferenceGenerator.Tests/Views/NamedControls.xml
