+++ src/XamlNameReferenceGenerator.Tests/Views/NoNamedControls.xml
