+++ src/XamlNameReferenceGenerator.Tests/Views/xNamedControl.xml
