+++ src/XamlNameReferenceGenerator.Tests/Views/xNamedControls.xml
