+++ src/XamlNameReferenceGenerator.Tests/XamlNameReferenceGenerator.Tests.csproj
@@ -11,13 +11,11 @@
         <PackageReference Include="xunit" Version="2.4.1" />
         <PackageReference Include="xunit.runner.console" Version="2.4.1" />
         <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3" />
-    </ItemGroup>
-    <ItemGroup>
         <PackageReference Include="Avalonia" Version="0.10.0-preview5" />
         <PackageReference Include="Avalonia.ReactiveUI" Version="0.10.0-preview5" />
     </ItemGroup>
     <ItemGroup>
-        <EmbeddedResource Include="Samples/*.xml" />
+        <EmbeddedResource Include="Views/*.xml" />
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="../XamlNameReferenceGenerator/XamlNameReferenceGenerator.csproj" />
